Fix max32664D funcs
