+++ app/CMakeLists.txt
@@ -22,7 +22,7 @@ FILE(GLOB app_sources   src/*.c
                         src/ui/screens/*.c
                         src/sm/*.c
                         )    
-                        
+
 target_sources(app PRIVATE ${app_sources})
 
 zephyr_library_include_directories(${CMAKE_CURRENT_SOURCE_DIR}/drivers)
@@ -35,6 +35,19 @@ set(gen_dir ${ZEPHYR_BINARY_DIR}/include/generated/)
 #zephyr_code_relocate(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/hw_module.c LOCATION EXTFLASH_TEXT NOCOPY)
 zephyr_code_relocate(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/images/*.c LOCATION EXTFLASH_TEXT NOCOPY)
 zephyr_code_relocate(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/fonts/*.c LOCATION EXTFLASH_TEXT NOCOPY)
+zephyr_code_relocate(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/sm/*.c LOCATION EXTFLASH_TEXT NOCOPY)
+zephyr_code_relocate(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/components/*.c LOCATION EXTFLASH_TEXT NOCOPY)
+zephyr_code_relocate(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/data_module.c LOCATION EXTFLASH_TEXT NOCOPY)
+
+
+zephyr_code_relocate(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/screens/*.c LOCATION EXTFLASH_TEXT NOCOPY)
+#zephyr_code_relocate(FILES ${WEST_TOPDIR}/application/drivers/sensor/max32664dmax32664d_bl.c LOCATION EXTFLASH_TEXT NOCOPY)
+#zephyr_code_relocate(FILES ${WEST_TOPDIR}/application/drivers/sensor/max32664d/max32664d_msbl.c LOCATION EXTFLASH_TEXT NOCOPY)
+#zephyr_code_relocate(FILES ${WEST_TOPDIR}/application/drivers/sensor/max32664c/*.* LOCATION EXTFLASH_TEXT NOCOPY)
+#zephyr_code_relocate(FILES ${WEST_TOPDIR}/application/drivers/sensor/max32664d/*.* LOCATION EXTFLASH_TEXT NOCOPY)
+#zephyr_code_relocate(FILES ${WEST_TOPDIR}/application/drivers/sensor/bmi323hpi/*.* LOCATION EXTFLASH_TEXT NOCOPY)
+zephyr_code_relocate(FILES $${WEST_TOPDIR}/application/drivers/sensor/max32664d/max32664d_bl.c LOCATION EXTFLASH_TEXT NOCOPY)
+zephyr_code_relocate(FILES $${WEST_TOPDIR}/application/drivers/sensor/max32664d/max32664d_msbl.h` LOCATION EXTFLASH_TEXT NOCOPY)
 
 zephyr_code_relocate(LIBRARY subsys__mgmt__mcumgr__mgmt LOCATION EXTFLASH_TEXT NOCOPY)
 zephyr_code_relocate(LIBRARY subsys__mgmt__mcumgr__transport LOCATION EXTFLASH_TEXT NOCOPY)
@@ -42,3 +55,5 @@ zephyr_code_relocate(LIBRARY subsys__mgmt__mcumgr__util LOCATION EXTFLASH_TEXT N
 zephyr_code_relocate(LIBRARY subsys__mgmt__mcumgr__smp LOCATION EXTFLASH_TEXT NOCOPY)
 zephyr_code_relocate(LIBRARY subsys__mgmt__mcumgr__grp__img_mgmt LOCATION EXTFLASH_TEXT NOCOPY)
 zephyr_code_relocate(LIBRARY subsys__mgmt__mcumgr__grp__os_mgmt LOCATION EXTFLASH_TEXT NOCOPY)
+
+zephyr_code_relocate(LIBRARY modules__cmsis-dsp LOCATION EXTFLASH_TEXT NOCOPY)
