+++ app/prj.conf
@@ -66,8 +66,8 @@ CONFIG_FILE_SYSTEM_LITTLEFS=y
 CONFIG_MPU_ALLOW_FLASH_WRITE=y
 
 # Enabled settings module and FS to store settings
-CONFIG_SETTINGS=y
-CONFIG_SETTINGS_RUNTIME=y
+CONFIG_SETTINGS=n
+CONFIG_SETTINGS_RUNTIME=n
 CONFIG_SETTINGS_FILE=y
 CONFIG_SETTINGS_FILE_PATH="/lfs/settings"
