+++ drivers/sensor/max32664d/max32664d.c
@@ -18,8 +18,6 @@
 
 LOG_MODULE_REGISTER(MAX32664D, CONFIG_MAX32664D_LOG_LEVEL);
 
-
-
 // #define DEFAULT_MODE_BPT_ESTIMATION 1
 #define MODE_RAW_SENSOR_DATA 2
 
@@ -78,12 +76,10 @@ uint8_t m_bpt_cal_vector[CALIBVECTOR_SIZE] = {0x50, 0x04, 0x03, 0, 0, 175, 63, 3
 											  207, 0, 4, 0, 3, 176, 22, 3, 33, 165, 0, 0, 0, 0, 15, 200, 2, 100, 3, 32, 0,
 											  0, 3, 207, 0, 4, 0, 3, 176, 102, 3};
 
-static int max32664_do_enter_app(const struct device *dev);
-
 static int m_read_op_mode(const struct device *dev)
 {
 	// struct max32664d_data *data = dev->data;
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t rd_buf[2] = {0x00, 0x00};
 
 	uint8_t wr_buf[2] = {0x02, 0x00};
@@ -104,7 +100,7 @@ static int m_read_op_mode(const struct device *dev)
 
 static int m_read_mcu_id(const struct device *dev)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t rd_buf[2] = {0x00, 0x00};
 	uint8_t wr_buf[2] = {0xFF, 0x00};
 
@@ -128,7 +124,7 @@ uint8_t max32664d_read_hub_status(const struct device *dev)
 	Field Reserved HostAccelUfInt FifoInOverInt FifoOutOvrInt DataRdyInt Err2 Err1 Err0
 	*/
 	// struct max32664_data *data = dev->data;
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t rd_buf[2] = {0x00, 0x00};
 	uint8_t wr_buf[2] = {0x00, 0x00};
 
@@ -147,7 +143,7 @@ uint8_t max32664d_read_hub_status(const struct device *dev)
 
 static int m_get_ver(const struct device *dev, uint8_t *ver_buf)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 
 	uint8_t wr_buf[2] = {0xFF, 0x03};
 
@@ -168,11 +164,9 @@ static int m_get_ver(const struct device *dev, uint8_t *ver_buf)
 	return 0;
 }
 
-
-
 static int m_i2c_write_cmd_4(const struct device *dev, uint8_t byte1, uint8_t byte2, uint8_t byte3, uint8_t byte4)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t wr_buf[4];
 	uint8_t rd_buf[1];
 
@@ -200,7 +194,7 @@ static int m_i2c_write_cmd_4(const struct device *dev, uint8_t byte1, uint8_t by
 
 static int m_i2c_write_cmd_5(const struct device *dev, uint8_t byte1, uint8_t byte2, uint8_t byte3, uint8_t byte4, uint8_t byte5)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t wr_buf[5];
 	uint8_t rd_buf[1];
 
@@ -229,7 +223,7 @@ static int m_i2c_write_cmd_5(const struct device *dev, uint8_t byte1, uint8_t by
 
 int max32664d_get_fifo_count(const struct device *dev)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t rd_buf[2] = {0x00, 0x00};
 	uint8_t wr_buf[2] = {0x12, 0x00};
 
@@ -248,7 +242,7 @@ int max32664d_get_fifo_count(const struct device *dev)
 
 static int m_i2c_write_cmd_3(const struct device *dev, uint8_t byte1, uint8_t byte2, uint8_t byte3, uint16_t cmd_delay)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t wr_buf[3];
 
 	uint8_t rd_buf[1] = {0x00};
@@ -277,7 +271,7 @@ static int m_i2c_write_cmd_3(const struct device *dev, uint8_t byte1, uint8_t by
 
 static int m_i2c_write_cmd_3_rsp_2(const struct device *dev, uint8_t byte1, uint8_t byte2, uint8_t byte3)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t wr_buf[3];
 
 	uint8_t rd_buf[2] = {0x00, 0x00};
@@ -308,7 +302,7 @@ static int m_i2c_write_cmd_3_rsp_2(const struct device *dev, uint8_t byte1, uint
 
 static int m_i2c_write_cmd_3_rsp_3(const struct device *dev, uint8_t byte1, uint8_t byte2, uint8_t byte3, uint8_t *rsp)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t wr_buf[3];
 
 	uint8_t rd_buf[3] = {0x00, 0x00, 0x00};
@@ -341,7 +335,7 @@ static int m_i2c_write_cmd_3_rsp_3(const struct device *dev, uint8_t byte1, uint
 
 static int m_i2c_write(const struct device *dev, uint8_t *wr_buf, uint32_t wr_len)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 
 	uint8_t rd_buf[1] = {0x00};
 
@@ -364,9 +358,9 @@ static int m_i2c_write(const struct device *dev, uint8_t *wr_buf, uint32_t wr_le
 	return 0;
 }
 
-static int max32664_do_enter_app(const struct device *dev)
+int max32664d_do_enter_app(const struct device *dev)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 
 	LOG_DBG("Entering app mode");
 
@@ -635,7 +629,7 @@ static int max32664_stop_estimation(const struct device *dev)
 int max32664_get_sample_fifo(const struct device *dev)
 {
 	struct max32664d_data *data = dev->data;
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 
 	uint8_t wr_buf[2] = {0x12, 0x01};
 
@@ -894,7 +888,7 @@ static const struct sensor_driver_api max32664_driver_api = {
 
 static int max32664_chip_init(const struct device *dev)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	// struct max32664d_data *data = dev->data;
 
 	if (!device_is_ready(config->i2c.bus))
@@ -906,7 +900,7 @@ static int max32664_chip_init(const struct device *dev)
 	gpio_pin_configure_dt(&config->reset_gpio, GPIO_OUTPUT);
 	gpio_pin_configure_dt(&config->mfio_gpio, GPIO_OUTPUT);
 
-	return max32664_do_enter_app(dev);
+	return max32664d_do_enter_app(dev);
 }
 
 #ifdef CONFIG_PM_DEVICE
@@ -939,7 +933,7 @@ static int max32664_pm_action(const struct device *dev,
  */
 #define MAX32664_DEFINE(inst)                                       \
 	static struct max32664d_data max32664_data_##inst;              \
-	static const struct max32664_config max32664_config_##inst =    \
+	static const struct max32664d_config max32664d_config_##inst =  \
 		{                                                           \
 			.i2c = I2C_DT_SPEC_INST_GET(inst),                      \
 			.reset_gpio = GPIO_DT_SPEC_INST_GET(inst, reset_gpios), \
@@ -950,7 +944,7 @@ static int max32664_pm_action(const struct device *dev,
 								 max32664_chip_init,                \
 								 PM_DEVICE_DT_INST_GET(inst),       \
 								 &max32664_data_##inst,             \
-								 &max32664_config_##inst,           \
+								 &max32664d_config_##inst,          \
 								 POST_KERNEL,                       \
 								 CONFIG_SENSOR_INIT_PRIORITY,       \
 								 &max32664_driver_api)
