+++ drivers/sensor/max32664d/max32664d_bl.c
@@ -3,12 +3,10 @@
 #include <zephyr/logging/log.h>
 #include <zephyr/pm/device.h>
 #include <zephyr/drivers/gpio.h>
-
-// #include "maxm86146_msbl.h"
-//  #include "max32664c_msbl.h"
-//#include "max32664d_msbl.h"
-
 #include "max32664d.h"
+
+//extern const uint8_t *max32664d_msbl;
+#include "max32664d_msbl.h"
 
 LOG_MODULE_REGISTER(MAX32664_BL, CONFIG_MAX32664D_LOG_LEVEL);
 
@@ -18,11 +16,11 @@ uint8_t max32664_fw_auth_vector[16] = {0};
 #define MAX32664_FW_UPDATE_WRITE_SIZE 8208 // Page size 8192 + 16 bytes for CRC
 #define MAX32664_FW_UPDATE_START_ADDR 0x4C
 
-#define MAX32664D_FW_BIN_INCLUDE 0
+#define MAX32664D_FW_BIN_INCLUDE 1
 
 static int m_read_bl_ver(const struct device *dev)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t rd_buf[4] = {0x00, 0x00, 0x00, 0x00};
 	uint8_t wr_buf[2] = {0x81, 0x00};
 
@@ -40,7 +38,7 @@ static int m_read_bl_ver(const struct device *dev)
 
 static int m_read_op_mode(const struct device *dev)
 { 
-    const struct max32664_config *config = dev->config;
+    const struct max32664d_config *config = dev->config;
     uint8_t rd_buf[2] = {0x00, 0x00};
 
     uint8_t wr_buf[2] = {0x02, 0x00};
@@ -61,7 +59,7 @@ static int m_read_op_mode(const struct device *dev)
 
 static int m_read_bl_page_size(const struct device *dev, uint16_t *bl_page_size)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t rd_buf[2] = {0x00, 0x00};
 	uint8_t wr_buf[2] = {0x81, 0x01};
 
@@ -80,7 +78,7 @@ static int m_read_bl_page_size(const struct device *dev, uint16_t *bl_page_size)
 
 static int m_write_set_num_pages(const struct device *dev, uint8_t num_pages)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t wr_buf[4] = {0x80, 0x02, 0x00, 0x00};
 	uint8_t rd_buf[1] = {0x00};
 
@@ -98,7 +96,7 @@ static int m_write_set_num_pages(const struct device *dev, uint8_t num_pages)
 
 static int m_write_init_vector(const struct device *dev, uint8_t *init_vector)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t wr_buf[13];
 	uint8_t rd_buf[1] = {0x00};
 
@@ -119,7 +117,7 @@ static int m_write_init_vector(const struct device *dev, uint8_t *init_vector)
 
 static int m_write_auth_vector(const struct device *dev, uint8_t *auth_vector)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t wr_buf[18];
 	uint8_t rd_buf[1] = {0x00};
 
@@ -144,7 +142,7 @@ volatile uint8_t tmp_wr_buf[8][1026];
 
 static int m_fw_write_page(const struct device *dev, uint8_t *msbl_data, uint32_t msbl_page_offset)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 
 	uint8_t rd_buf[1] = {0x00};
 	uint8_t cmd_wr_buf[2] = {0x80, 0x04};
@@ -201,7 +199,7 @@ static int m_fw_write_page(const struct device *dev, uint8_t *msbl_data, uint32_
 
 static int m_erase_app(const struct device *dev)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 	uint8_t wr_buf[2] = {0x80, 0x03};
 	uint8_t rd_buf[1] = {0x00};
 
@@ -214,6 +212,25 @@ static int m_erase_app(const struct device *dev)
 
 	LOG_DBG("Erase App : RSP: %x", rd_buf[0]);
 }
+
+static int m_read_mcu_id(const struct device *dev)
+{
+	const struct max32664d_config *config = dev->config;
+	uint8_t rd_buf[2] = {0x00, 0x00};
+	uint8_t wr_buf[2] = {0xFF, 0x00};
+
+	k_sleep(K_USEC(300));
+	i2c_write_dt(&config->i2c, wr_buf, sizeof(wr_buf));
+	k_sleep(K_MSEC(MAX32664_DEFAULT_CMD_DELAY));
+
+	i2c_read_dt(&config->i2c, rd_buf, sizeof(rd_buf));
+	k_sleep(K_MSEC(MAX32664_DEFAULT_CMD_DELAY));
+
+	printk("MCU ID = %x %x\n", rd_buf[0], rd_buf[1]);
+
+	return 0;
+}
+
 
 static int max32664d_load_fw(const struct device *dev, uint8_t *fw_bin_array)
 {
@@ -256,7 +273,7 @@ static int max32664d_load_fw(const struct device *dev, uint8_t *fw_bin_array)
 	}
 #endif
 
-	max32664_do_enter_app(dev);
+	max32664d_do_enter_app(dev);
 
 	printk("End Load MSBL\n---\n");
 
@@ -266,7 +283,7 @@ static int max32664d_load_fw(const struct device *dev, uint8_t *fw_bin_array)
 
 void max32664_do_enter_bl(const struct device *dev)
 {
-	const struct max32664_config *config = dev->config;
+	const struct max32664d_config *config = dev->config;
 
 	LOG_INF("Entering Bootloader mode\n");
